+++ src/life-cycle/process-html-meta.ts
@@ -69,6 +69,13 @@ export async function processHtmlMetaRefresh(
         linkType, depth, link, refUrl,
         res.localRoot, options.encoding[linkType],
         savePath, res.type);
+      if (!resource) {
+        if (skip.isTraceEnabled()) {
+          skip.trace('skip generateSavePath',
+            originalLink, link, linkType, refUrl);
+        }
+        continue;
+      }
       resource = await pipeline.processBeforeDownload(resource, elem, res, options);
       if (!resource) {
         if (skip.isTraceEnabled()) {