+++ src/life-cycle/process-svg.ts
@@ -67,6 +67,13 @@ export async function processSvg(
         linkType, depth, link, refUrl,
         res.localRoot, options.encoding[linkType],
         savePath, res.type);
+      if (!resource) {
+        if (skip.isTraceEnabled()) {
+          skip.trace('skip generateSavePath',
+            originalLink, link, linkType, refUrl);
+        }
+        continue;
+      }
       resource = await pipeline.processBeforeDownload(resource, elem, res, options);
       if (!resource) {
         if (skip.isTraceEnabled()) {
@@ -93,4 +100,3 @@ export async function processSvg(
   res.body = getResourceBodyFromHtml(res, options);
   return res;
 }
-