+++ src/life-cycle/save-resource-to-disk.ts
@@ -29,12 +29,15 @@ export async function saveResourceToDisk(
       res.depth, res.url, res.redirectedUrl, localRoot,
       res.encoding, undefined, res.type);
     // maybe we can try module:fs/promises.symlink first
-    if (redirectResource.replacePath) {
+    if (redirectResource?.replacePath) {
       const savePath = decodeURI(res.savePath);
       await writeFile(path.join(localRoot, savePath), res.body, res.encoding, mtime);
       const redirectedResource = await pipeline.createResource(res.type,
         res.depth, res.redirectedUrl, res.refUrl, res.localRoot,
         res.encoding, undefined, res.type);
+      if (!redirectedResource) {
+        return;
+      }
       const redirectedSavePath = decodeURI(redirectedResource.savePath);
       await writeFile(path.join(localRoot, redirectedSavePath), res.body,
         res.encoding, mtime);