+++ test/life-cycle/existing-resource.spec.ts
@@ -70,6 +70,7 @@ function makeLifeCycle(
     init: [],
     linkRedirect: [],
     detectResourceType: [],
+    generateSavePath: [],
     createResource,
     processBeforeDownload: [],
     download: [