+++ test/life-cycle/init-submit.spec.ts
@@ -36,6 +36,7 @@ function makeLifeCycle(
     init,
     linkRedirect: [],
     detectResourceType: [],
+    generateSavePath: [],
     createResource,
     processBeforeDownload: [],
     download: [],