+++ test/life-cycle/process-html.spec.ts
@@ -19,6 +19,7 @@ const testLifeCycle = (): ProcessingLifeCycle => ({
   init: [],
   linkRedirect: [],
   detectResourceType: [],
+  generateSavePath: [],
   createResource,
   processBeforeDownload: [],
   download: [